Deployment Tool: Remove create_venv + fix --dry-run/--init
- As the deployment tool's code base grows larger, I realized that
  the ability to create a venv using pyside6-deploy was a bit of
  over engineering. There are instances where I have to use the
  current Python interpreter to fetch some information from the
  newly created venv Python, which results in weird code.

  Note: The tool would still work even if the user is using a
  globally installed Python interpreter with PySide6 installed.

- Now, the user is warned if he is not in a virtual environment and
  prompted the requirement to install further Python packages. If
  the user input's "no", then the tool exits.

- dry_run used to create an empty 'pysidedeploy.spec' which can
  wreck the normal deployment process. This is fixed by
  pyside6-deploy using the 'default.spec' instead of creating
  a new 'pysidedeploy.spec'.

Fixes: PYSIDE-2258
Task-number: PYSIDE-1612
Pick-to: 6.4 6.5

+++ sources/pyside-tools/deploy_lib/config.py
@@ -24,8 +24,11 @@ class Config:
         self.config_file = config_file
         self.parser = ConfigParser(comment_prefixes="/", allow_no_value=True)
         if not self.config_file.exists():
-            logging.info(f"[DEPLOY] Creating config file {self.config_file}")
-            shutil.copy(Path(__file__).parent / "default.spec", self.config_file)
+            if not dry_run:
+                logging.info(f"[DEPLOY] Creating config file {self.config_file}")
+                shutil.copy(Path(__file__).parent / "default.spec", self.config_file)
+            else:
+                self.config_file = Path(__file__).parent / "default.spec"
         else:
             print(f"Using existing config file {config_file}")
         self.parser.read(self.config_file)
